Move WebIDL constants machinery to bindings::constant
