+++ components/script/dom/bindings/mod.rs
@@ -131,6 +131,7 @@
 pub use style::domrefcell as cell;
 
 pub mod callback;
+pub mod constant;
 pub mod conversions;
 pub mod error;
 pub mod global;
